requirements.txt
aws-cdk-lib==2.139.1
constructs>=10.0.0,<11.0.0
aws-cdk-lambda-layer-kubectl-v29
pytest==8.2.0
moto==5.0.6
boto3>=1.34.98
